finalizado modulo de crud do departamento, iniciado crud de funcionario com filtro por pipe

+++ src/app/components/admin/departamento/departamento.component.ts
@@ -66,10 +66,12 @@ export class DepartamentoComponent implements OnInit {
       confirmButtonText: 'Sim',
       cancelButtonText: 'Não'
     }).then((result) => {
+      if (result.value){
       this.departamentoService.delete(depto.id)
         .then(()=> {
-          Swal.fire('Departamento excluído com sucessi!','','success')
+          Swal.fire('Departamento excluído com sucesso!','','success')
         })
+      }
     })
   }
 }
